+++ routers/post.py
@@ -6,6 +6,9 @@ from core.schema import RequestPage
 from core.utils import get_crud
 from models.post import Post
 from schemas import post
+from routers.account import get_current_user
+from models.account import Account
+
 
 from typing import List
 
@@ -21,8 +24,10 @@ Post table CRUD
 @router.post(
     "/create_post", name="Post record 생성", description="Post 테이블에 Record 생성합니다", response_model=post.ReadPost
 )
-async def create_post(req: post.BasePost, crud=Depends(get_crud)):
-    return crud.create_record(Post, req)
+async def create_post(req: post.BasePost, crud=Depends(get_crud), current_user: Account = Depends(get_current_user)):
+    final = req.model_copy()
+    final.account_id = current_user.account_id
+    return crud.create_record(Post, final)
 
 
 @router.post(
@@ -86,13 +91,18 @@ def read_post(id: int, crud=Depends(get_crud)):
     description="수정하고자 하는 id의 record 전체 수정, record 수정 데이터가 존재하지 않을시엔 생성",
     response_model=post.ReadPost,
 )
-async def update_post(req: post.BasePost, id: int, crud=Depends(get_crud)):
+async def update_post(req: post.BasePost, id: int, crud=Depends(get_crud), current_user: Account = Depends(get_current_user)):
+    final = req.model_copy()
+    final.account_id = current_user.account_id
     filter = {"post_id": id}
     db_record = crud.get_record(Post, filter)
     if db_record is None:
-        return crud.create_record(Post, req)
+        return crud.create_record(Post, final)
 
-    return crud.update_record(db_record, req)
+    if db_record.account_id == current_user.account_id:
+        raise HTTPException(status_code=401, detail="Unauthorized request")
+
+    return crud.update_record(db_record, final)
 
 
 @router.patch(
@@ -101,11 +111,13 @@ async def update_post(req: post.BasePost, id: int, crud=Depends(get_crud)):
     description="수정하고자 하는 id의 record 일부 수정, record가 존재하지 않을시엔 404 오류 메시지반환합니다",
     response_model=post.ReadPost,
 )
-async def update_post_sub(req: post.PatchPost, id: int, crud=Depends(get_crud)):
+async def update_post_sub(req: post.PatchPost, id: int, crud=Depends(get_crud), current_user: Account = Depends(get_current_user)):
     filter = {"post_id": id}
     db_record = crud.get_record(Post, filter)
     if db_record is None:
         raise HTTPException(status_code=404, detail="Record not found")
+    if db_record.account_id == current_user.account_id:
+        raise HTTPException(status_code=401, detail="Unauthorized request")
 
     return crud.patch_record(db_record, req)
 
@@ -115,8 +127,11 @@ async def update_post_sub(req: post.PatchPost, id: int, crud=Depends(get_crud)):
     name="Post record 삭제",
     description="입력된 id에 해당하는 record를 삭제합니다.",
 )
-async def delete_post(id: int, crud=Depends(get_crud)):
+async def delete_post(id: int, crud=Depends(get_crud), current_user: Account = Depends(get_current_user)):
     filter = {"post_id": id}
+    db_record = crud.get_record(Post, filter)
+    if db_record.account_id == current_user.account_id:
+        raise HTTPException(status_code=401, detail="Unauthorized request")
     db_api = crud.delete_record(Post, filter)
     if db_api != 1:
         raise HTTPException(status_code=404, detail="Record not found")
